updated: user profile response and service logic

+++ src/main/java/kh/edu/istad/codecompass/dto/problem/response/ProblemResponse.java
@@ -20,7 +20,7 @@ public record ProblemResponse(
         @JsonProperty("best_time_execution")
         Double bestTimeExecution,
 
-        Byte coin,
+        Integer coin,
 
         String description,
