+++ src/main/java/kh/edu/istad/codecompass/dto/problem/response/ProblemResponseBySpecificUser.java
@@ -17,7 +17,7 @@ public record ProblemResponseBySpecificUser (
         @JsonProperty("best_time_execution")
         Double bestTimeExecution,
 
-        Byte coin,
+        Integer coin,
 
         String description,
